fix(frontend): reset stale state on load, extract explorer helpers into util module
